+++ packages/chaingraph-nodes/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @badaitech/chaingraph-nodes
 
+## 0.1.28
+
+### Patch Changes
+
+- 0g llm node implementation and fixes
+- Updated dependencies
+  - @badaitech/chaingraph-types@0.1.28
+
 ## 0.1.27
 
 ### Patch Changes