+++ packages/typescript-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @badaitech/typescript-config
 
+## 0.1.28
+
+### Patch Changes
+
+- 0g llm node implementation and fixes
+
 ## 0.1.27
 
 ### Patch Changes